requirements.txt
streamlit
langgraph
langchain
langchain-google-genai
langchain-community
langchain-core
duckduckgo-search
beautifulsoup4
requests
python-dotenv